+++ uds/translator/data_record_definitions/__init__.py
@@ -8,7 +8,11 @@ from .conditional import (
     CONDITIONAL_CHALLENGE_SERVER,
     CONDITIONAL_DATA,
     CONDITIONAL_DATA_FROM_MEMORY,
+    CONDITIONAL_DIR_INFO,
     CONDITIONAL_EVENT_TYPE_RECORD_09_2020,
+    CONDITIONAL_FILE_AND_PATH_NAME,
+    CONDITIONAL_FILE_SIZES,
+    CONDITIONAL_FILE_SIZES_OR_DIR_INFO,
     CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH,
     CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH_FILE_TRANSFER,
     CONDITIONAL_MEMORY_ADDRESS_AND_SIZE,
@@ -17,7 +21,7 @@ from .conditional import (
     CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_CLIENT,
     CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_SERVER,
     CONDITIONAL_OPTIONAL_NEEDED_ADDITIONAL_PARAMETER,
-    CONDITIONAL_OPTIONAL_SESSION_KEY_INFO,SCALING_DATA_RECORDS,
+    CONDITIONAL_OPTIONAL_SESSION_KEY_INFO,
     CONDITIONAL_PROOF_OF_OWNERSHIP_CLIENT,
     CONDITIONAL_PROOF_OF_OWNERSHIP_SERVER,
     DTC_SNAPSHOT_RECORDS_LIST_2013,
@@ -31,6 +35,7 @@ from .conditional import (
     EVENT_TYPE_RECORD_07_2013,
     EVENT_TYPE_RECORD_07_2020,
     EVENT_TYPE_RECORD_09_2020,
+    SCALING_DATA_RECORDS,
 )
 from .did import (
     DATA_FROM_DID_2013,
@@ -99,10 +104,6 @@ from .other import (
     CHALLENGE_SERVER_LENGTH,
     COMMUNICATION_CONFIGURATION,
     CONDITIONAL_COMMUNICATION_CONTROL_REQUEST,
-    CONDITIONAL_DIR_INFO,
-    CONDITIONAL_FILE_AND_PATH_NAME,
-    CONDITIONAL_FILE_SIZES,
-    CONDITIONAL_FILE_SIZES_OR_DIR_INFO,
     CONDITIONAL_LINK_CONTROL_REQUEST,
     CONDITIONAL_POWER_DOWN_TIME,
     CONDITIONAL_SECURED_DATA_TRANSMISSION_REQUEST,
